Change score typr to decimal

+++ server/src/models/MCQStat.js
@@ -5,7 +5,7 @@ const MCQStat = db.define(
     "MCQStat",
     {
         score: {
-            type: Sequelize.INTEGER(3),
+            type: Sequelize.DECIMAL(4, 2),
             allowNull: false
         },
         applicantId: {
